Extract timeline runway into composable components
Replace the monolithic Scrubber component with a layered runway/
architecture where each component has a single, clear responsibility:

- RunwayViewport: perspective positioning/scaling when drawer opens
  (translateY/scaleY from PR #412)
- RunwayTilt: 3D perspective tilt (rotateX + foreshortening compensation)
- Playhead: playback position overlay (plasma canvas + resize tracking)
- Runway: top-level coordinator wiring geometry, scroll, and children

Split useScrubber into useRunwayGeometry (perspective math + container
tracking) and useRunwayScroll (scroll↔time sync, animation loop,
FrequencyVisualizer lifecycle).

Move ZoomControls and PlasmaPlayhead into the runway/ folder. Rename
CSS classes from scrubber__ to runway__ (viewport, tilt, playhead).
Update all e2e selectors accordingly.
