+++ src/components/workstation/runway/Runway.css
@@ -1,4 +1,4 @@
-.scrubber {
+.runway {
   flex: 1;
   display: flex;
   flex-direction: column;
@@ -8,7 +8,9 @@
   min-height: 0;
 }
 
-.scrubber__perspective {
+/* RunwayViewport: perspective positioning/scaling when drawer opens */
+
+.runway__viewport {
   flex: 1;
   display: flex;
   flex-direction: column;
@@ -18,13 +20,15 @@
 }
 
 @media (prefers-reduced-motion: reduce) {
-  .scrubber__perspective {
+  .runway__viewport {
     perspective: none;
   }
 }
 
-.scrubber__timeline {
-  /* Fill the perspective container. The inline transform tilts and scales
+/* RunwayTilt: 3D perspective tilt + scroll container */
+
+.runway__tilt {
+  /* Fill the viewport container. The inline transform tilts and scales
      the timeline for the runway perspective effect. */
   flex: 1;
   display: flex;
@@ -35,11 +39,13 @@
   /* Size container for cqh units in .timeline padding.
      The timeline padding must reference the scroll container's height
      (not the viewport) so the content boundary at time=0 aligns with
-     the cursor position (which is a % of the scrubber height). */
+     the cursor position (which is a % of the runway height). */
   container-type: size;
 }
 
-.scrubber__cursor {
+/* Playhead: playback position overlay */
+
+.runway__playhead {
   position: absolute;
   top: calc(
     var(--cursor-position) * (100% - var(--drawer-height, 0px)) - 120px
