+++ src/components/workstation/runway/__tests__/Runway.test.tsx
@@ -2,7 +2,7 @@ import { act, fireEvent, render } from '@testing-library/react';
 import { createRef } from 'react';
 import * as Tone from 'tone';
 import AudioService from '../../../../services/AudioService';
-import Scrubber, { type ScrubberHandle } from '../Scrubber';
+import Runway, { type RunwayHandle } from '../Runway';
 
 const audioService = AudioService.getInstance();
 const playbackService = audioService.playbackService;
@@ -25,37 +25,33 @@ afterEach(() => {
 it('pauses playback when timeline is scrolled while playing', () => {
   playbackService.play();
 
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
   fireEvent.scroll(timeline);
 
   expect(playbackService.isPlaying).toBe(false);
 });
 
 it('does not pause playback when timeline is scrolled while paused', () => {
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
   fireEvent.scroll(timeline);
 
   expect(playbackService.isPlaying).toBe(false);
 });
 
 it('positions perspective-origin and transform-origin at runway bottom', () => {
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  const perspective = container.querySelector(
-    '.scrubber__perspective',
-  ) as HTMLElement;
-  const timeline = container.querySelector(
-    '.scrubber__timeline',
-  ) as HTMLElement;
+  const viewport = container.querySelector('.runway__viewport') as HTMLElement;
+  const tilt = container.querySelector('.runway__tilt') as HTMLElement;
 
   // Both origins should use the same Y coordinate (runway bottom).
   // In jsdom offsetHeight is 0, so runwayBottomY = 0.75 * 0 = 0.
-  expect(perspective.style.perspectiveOrigin).toBe('center 0px');
-  expect(timeline.style.transformOrigin).toBe('center 0px');
+  expect(viewport.style.perspectiveOrigin).toBe('center 0px');
+  expect(tilt.style.transformOrigin).toBe('center 0px');
 });
 
 it('keeps perspective geometry stable when drawer height changes', () => {
@@ -71,18 +67,16 @@ it('keeps perspective geometry stable when drawer height changes', () => {
     },
   });
 
-  const { container, rerender } = render(<Scrubber {...defaultProps} />);
+  const { container, rerender } = render(<Runway {...defaultProps} />);
 
-  const timeline = container.querySelector(
-    '.scrubber__timeline',
-  ) as HTMLElement;
+  const tilt = container.querySelector('.runway__tilt') as HTMLElement;
 
-  const transformWithoutDrawer = timeline.style.transform;
+  const transformWithoutDrawer = tilt.style.transform;
 
   // Open the drawer — the perspective geometry should NOT change
-  rerender(<Scrubber {...defaultProps} drawerHeight={280} />);
+  rerender(<Runway {...defaultProps} drawerHeight={280} />);
 
-  const transformWithDrawer = timeline.style.transform;
+  const transformWithDrawer = tilt.style.transform;
 
   expect(transformWithDrawer).toBe(transformWithoutDrawer);
 
@@ -96,7 +90,7 @@ it('keeps perspective geometry stable when drawer height changes', () => {
   }
 });
 
-it('applies translateY and scaleY to perspective div when drawer is open', () => {
+it('applies translateY and scaleY to viewport div when drawer is open', () => {
   const containerHeight = 800;
   const originalDescriptor = Object.getOwnPropertyDescriptor(
     HTMLElement.prototype,
@@ -111,18 +105,16 @@ it('applies translateY and scaleY to perspective div when drawer is open', () =>
 
   const drawerHeight = 280;
   const { container } = render(
-    <Scrubber {...defaultProps} drawerHeight={drawerHeight} />,
+    <Runway {...defaultProps} drawerHeight={drawerHeight} />,
   );
 
-  const perspective = container.querySelector(
-    '.scrubber__perspective',
-  ) as HTMLElement;
+  const viewport = container.querySelector('.runway__viewport') as HTMLElement;
 
   const visibleHeight = containerHeight - drawerHeight;
   const expectedScaleY = visibleHeight / containerHeight;
   const expectedTranslateY = -drawerHeight / 2;
 
-  expect(perspective.style.transform).toBe(
+  expect(viewport.style.transform).toBe(
     `translateY(${expectedTranslateY}px) scaleY(${expectedScaleY})`,
   );
 
@@ -135,48 +127,46 @@ it('applies translateY and scaleY to perspective div when drawer is open', () =>
   }
 });
 
-it('does not apply perspective transform when drawer is closed', () => {
-  const { container } = render(<Scrubber {...defaultProps} drawerHeight={0} />);
+it('does not apply viewport transform when drawer is closed', () => {
+  const { container } = render(<Runway {...defaultProps} drawerHeight={0} />);
 
-  const perspective = container.querySelector(
-    '.scrubber__perspective',
-  ) as HTMLElement;
+  const viewport = container.querySelector('.runway__viewport') as HTMLElement;
 
-  expect(perspective.style.transform).toBe('');
+  expect(viewport.style.transform).toBe('');
 });
 
 it('does not render a shade overlay', () => {
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  expect(container.querySelector('.scrubber__shade')).toBeNull();
+  expect(container.querySelector('.runway__shade')).toBeNull();
 });
 
-it('passes drawer height as CSS variable to cursor for position alignment', () => {
+it('passes drawer height as CSS variable to playhead for position alignment', () => {
   const drawerHeight = 200;
 
   const { container } = render(
-    <Scrubber {...{ ...defaultProps, drawerHeight }} />,
+    <Runway {...{ ...defaultProps, drawerHeight }} />,
   );
 
-  const cursor = container.querySelector('.scrubber__cursor') as HTMLElement;
+  const playhead = container.querySelector('.runway__playhead') as HTMLElement;
 
-  // The cursor element must expose --drawer-height so CSS can position the
+  // The playhead element must expose --drawer-height so CSS can position the
   // playhead within the visible area above the drawer.
-  const heightVar = cursor.style.getPropertyValue('--drawer-height');
+  const heightVar = playhead.style.getPropertyValue('--drawer-height');
   expect(heightVar).toBe(`${drawerHeight}px`);
-  expect(cursor.style.transform).not.toContain('scaleY');
+  expect(playhead.style.transform).not.toContain('scaleY');
 });
 
-it('perspective wrapper handles wheel events for full hit-area coverage', () => {
+it('viewport wrapper handles wheel events for full hit-area coverage', () => {
   playbackService.play();
 
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  // The perspective wrapper covers the full rectangular area. Wheel events
+  // The viewport wrapper covers the full rectangular area. Wheel events
   // landing outside the tilted scroll container's trapezoid hit the wrapper
   // instead, which forwards them as programmatic scrolls.
-  const perspectiveWrapper = container.querySelector('.scrubber__perspective')!;
-  fireEvent.wheel(perspectiveWrapper, { deltaY: 100 });
+  const viewport = container.querySelector('.runway__viewport')!;
+  fireEvent.wheel(viewport, { deltaY: 100 });
 
   expect(playbackService.isPlaying).toBe(false);
 });
@@ -195,7 +185,7 @@ it('does not stop playback at end of scroll during recording', () => {
   recordingService.arm();
   recordingService.startRecording();
 
-  render(<Scrubber {...defaultProps} />);
+  render(<Runway {...defaultProps} />);
 
   // In jsdom, scrollWidth equals clientWidth (no overflow), so the
   // end-of-scroll condition is satisfied. During recording this must NOT
@@ -215,10 +205,10 @@ it('stops recording when timeline is clicked during recording', () => {
   const onStopRecording = vi.fn();
 
   const { container } = render(
-    <Scrubber {...defaultProps} onStopRecording={onStopRecording} />,
+    <Runway {...defaultProps} onStopRecording={onStopRecording} />,
   );
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
   fireEvent.click(timeline);
 
   expect(onStopRecording).toHaveBeenCalledOnce();
@@ -233,10 +223,10 @@ it('cancels count-in when timeline is clicked during count-in', () => {
   const onStopRecording = vi.fn();
 
   const { container } = render(
-    <Scrubber {...defaultProps} onStopRecording={onStopRecording} />,
+    <Runway {...defaultProps} onStopRecording={onStopRecording} />,
   );
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
   fireEvent.click(timeline);
 
   expect(onStopRecording).toHaveBeenCalledOnce();
@@ -248,9 +238,9 @@ it('does not pause playback when timeline is scrolled during recording', () => {
   recordingService.arm();
   recordingService.startRecording();
 
-  const { container } = render(<Scrubber {...defaultProps} />);
+  const { container } = render(<Runway {...defaultProps} />);
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
   fireEvent.scroll(timeline);
 
   expect(playbackService.isPlaying).toBe(true);
@@ -272,7 +262,7 @@ it('does not update transportTime during count-in', () => {
   recordingService.startRecording();
   recordingService.startCountIn();
 
-  render(<Scrubber {...defaultProps} />);
+  render(<Runway {...defaultProps} />);
 
   act(() => {
     rafCallback(0);
@@ -284,11 +274,11 @@ it('does not update transportTime during count-in', () => {
 });
 
 it('syncs timeline scroll position via imperative handle (inverted scroll)', () => {
-  const ref = createRef<ScrubberHandle>();
+  const ref = createRef<RunwayHandle>();
 
-  const { container } = render(<Scrubber ref={ref} {...defaultProps} />);
+  const { container } = render(<Runway ref={ref} {...defaultProps} />);
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
 
   // Mock scroll dimensions so maxScrollTop is non-zero
   Object.defineProperty(timeline, 'scrollHeight', {
@@ -311,11 +301,11 @@ it('syncs timeline scroll position via imperative handle (inverted scroll)', () 
 });
 
 it('scrolls to maxScrollTop when time is zero (beginning at bottom)', () => {
-  const ref = createRef<ScrubberHandle>();
+  const ref = createRef<RunwayHandle>();
 
-  const { container } = render(<Scrubber ref={ref} {...defaultProps} />);
+  const { container } = render(<Runway ref={ref} {...defaultProps} />);
 
-  const timeline = container.querySelector('.scrubber__timeline')!;
+  const timeline = container.querySelector('.runway__tilt')!;
 
   Object.defineProperty(timeline, 'scrollHeight', {
     value: 2000,
